requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pyjwt
passlib[bcrypt]
python-dotenv
httpx
slowapi
pytest
pytest-asyncio
